requirements.txt
discord.py
asyncpg
aiohttp
selenium
beautifulsoup4
python-dotenv
